Fixed file structure of test utils
